refactor(elevation): rename everything from elevation to elevation-service
