+++ examples/elevation-service/components/map-canvas/map-canvas.tsx
