+++ library/src/index.ts
@@ -2,7 +2,7 @@
 export * from './hooks/autocomplete';
 export * from './hooks/directions-service';
 export * from './hooks/distance-matrix-service';
-export * from './hooks/elevation';
+export * from './hooks/elevation-service';
 export * from './hooks/geocoder';
 export * from './hooks/map-instance';
 export * from './hooks/max-zoom';
