+++ package.json
@@ -17,7 +17,7 @@
     "start:places-autocomplete-example": "EXAMPLE=places-autocomplete run-p start:library start:example",
     "start:directions-service-example": "EXAMPLE=directions-service run-p start:library start:example",
     "start:distance-matrix-service-example": "EXAMPLE=distance-matrix-service run-p start:library start:example",
-    "start:elevation-example": "EXAMPLE=elevation run-p start:library start:example",
+    "start:elevation-service-example": "EXAMPLE=elevation-service run-p start:library start:example",
     "preversion": "echo \"To create a new library version run 'npm version -w library' in the repository root.\""
   },
   "keywords": [
